+++ tests/test_pendo_syncher.py
@@ -11,7 +11,7 @@ from tests.helpers.test_utils import MockResponseObject
 
 
 def test_pendo_syncher_request_body(mocker, db_create_host):
-    mock_request_function_mock = mocker.patch("lib.pendo_sync_helpers._make_request")
+    mock_request_function_mock = mocker.patch("pendo_syncher._make_request")
 
     host = db_create_host()
 
@@ -35,13 +35,13 @@ def test_pendo_syncher_response_process(mocker, db_create_host):
     mock_response.status_code = 200
     mock_response.content = json.dumps(response_content)
 
-    request_session_post_mock = mocker.patch("lib.pendo_sync_helpers.Session.post")
+    request_session_post_mock = mocker.patch("pendo_syncher.Session.post")
     request_session_post_mock.side_effect = mock_response
 
     config = Config(RuntimeEnvironment.PENDO_JOB)
     config.pendo_sync_active = True
 
-    mock_pendo_failure = mocker.patch("lib.pendo_sync_helpers.pendo_failure")
+    mock_pendo_failure = mocker.patch("pendo_syncher.pendo_failure")
     pendo_syncher_run(config, mock.Mock(), db.session, shutdown_handler=mock.Mock(**{"shut_down.return_value": False}))
 
     mock_pendo_failure.assert_not_called()
@@ -53,20 +53,20 @@ def test_pendo_syncher_response_process_failure(mocker, db_create_host):
     mock_response = MockResponseObject()
     mock_response.status_code = 403
 
-    request_session_post_mock = mocker.patch("lib.pendo_sync_helpers.Session.post")
+    request_session_post_mock = mocker.patch("pendo_syncher.Session.post")
     request_session_post_mock.side_effect = mock_response
 
     config = Config(RuntimeEnvironment.PENDO_JOB)
     config.pendo_sync_active = True
 
-    mock_pendo_failure = mocker.patch("lib.pendo_sync_helpers.pendo_failure")
+    mock_pendo_failure = mocker.patch("pendo_syncher.pendo_failure")
     pendo_syncher_run(config, mock.Mock(), db.session, shutdown_handler=mock.Mock(**{"shut_down.return_value": False}))
 
     mock_pendo_failure.assert_called_once()
 
 
 def test_pendo_syncher_exception(mocker, db_create_host):
-    request_session_post_mock = mocker.patch("lib.pendo_sync_helpers.Session.post")
+    request_session_post_mock = mocker.patch("pendo_syncher.Session.post")
     request_session_post_mock.side_effect = Exception()
 
     db_create_host()
@@ -74,14 +74,14 @@ def test_pendo_syncher_exception(mocker, db_create_host):
     config = Config(RuntimeEnvironment.PENDO_JOB)
     config.pendo_sync_active = True
 
-    mock_pendo_failure = mocker.patch("lib.pendo_sync_helpers.pendo_failure")
+    mock_pendo_failure = mocker.patch("pendo_syncher.pendo_failure")
     pendo_syncher_run(config, mock.Mock(), db.session, shutdown_handler=mock.Mock(**{"shut_down.return_value": False}))
 
     mock_pendo_failure.assert_called_once()
 
 
 def test_pendo_syncher_retry_error(mocker, db_create_host):
-    request_session_post_mock = mocker.patch("lib.pendo_sync_helpers.Session.post")
+    request_session_post_mock = mocker.patch("pendo_syncher.Session.post")
     request_session_post_mock.side_effect = exceptions.RetryError
 
     db_create_host()
@@ -89,7 +89,7 @@ def test_pendo_syncher_retry_error(mocker, db_create_host):
     config = Config(RuntimeEnvironment.PENDO_JOB)
     config.pendo_sync_active = True
 
-    mock_pendo_failure = mocker.patch("lib.pendo_sync_helpers.pendo_failure")
+    mock_pendo_failure = mocker.patch("pendo_syncher.pendo_failure")
     pendo_syncher_run(config, mock.Mock(), db.session, shutdown_handler=mock.Mock(**{"shut_down.return_value": False}))
 
     mock_pendo_failure.assert_called_once()
